backend/requirements/dev.txt
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pyjwt==2.9.0
asgiref==3.8.1
sqlparse==0.5.1
tzdata==2024.2
python-dotenv==1.0.1
psycopg2_binary==2.9.10
pytz
django-cors-headers==4.6.0
